# docs/swagger.yaml
definitions:
  models.Credentials:
    properties:
      email:
        example: emirkovacevic@protonmail.com
        type: string
      password:
        example: password123As!
        type: string
    required:
    - email
    - password
    type: object
  models.ItemNew:
    properties:
      description:
        example: My Item Description
        type: string
      name:
        example: My Item
        type: string
      price:
        $ref: '#/definitions/models.Price'
      quantity:
        example: 123
        type: integer
    required:
    - description
    - name
    - price
    - quantity
    type: object
...
  models.Price:
    properties:
      amount:
        example: 123
        type: integer
      currency:
        example: USD
        type: string
    required:
    - amount
    - currency
    type: object
...
  responses.AuthResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      message:
        type: string
      status:
        type: integer
    type: object
  responses.ItemResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.AuthResponse'
      summary: Log in a user
      tags:
      - Auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Logs out a user and clears the session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.AuthResponse'
      summary: Logout a user
      tags:
      - Auth
  /api/item/{itemId}:
    get:
      consumes:
      - application/json
      description: Gets an item
      parameters:
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ItemResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ItemResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ItemResponse'
      security:
      - BearerAuth: []
      summary: Get an item
      tags:
      - Item
  /api/item/{organisationId}:
    post:
      consumes:
      - application/json
      description: Creates an item
      parameters:
      - description: Organisation ID
        in: path
        name: organisationId
        required: true
        type: string
      - description: Item data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.ItemNew'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ItemResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ItemResponse'
      security:
      - BearerAuth: []
      summary: Create an item
      tags:
      - Item
  /api/items/{organisationId}:
    get:
      consumes:
      - application/json
      description: Gets items
      parameters:
      - description: Organisation ID
        in: path
        name: organisationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ItemResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ItemResponse'
      security:
      - BearerAuth: []
      summary: Get items
      tags:
      - Item
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"